requirements.txt
playwright>=1.44.0
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
feedparser>=6.0.0
praw>=7.7.0
googlesearch-python>=1.2.0
colorlog>=6.8.0
tqdm>=4.66.0
